client_requirements.txt
aiohttp==3.9.5
cryptography==46.0.3
python-dotenv==1.2.1
websocket_client==1.8.0
